Added shooting to towers, added 2nd tower
Added flexible system for shooting (only used by the towers), using an IProjectile interface. In that way, I can use the code for 2 different towers (one that shoots straight ahead and other that shoots in an arc).
Also added a health system, so that I can use it to deal damage to the enemies and the towers